Rename packages and rename Main class
