+++ backend/src/main/java/com/curevent/CureventApplication.java
@@ -1,13 +1,13 @@
-package com.curevent.demo;
+package com.curevent;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class DemoApplication {
+public class CureventApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(DemoApplication.class, args);
+		SpringApplication.run(CureventApplication.class, args);
 	}
 
 }
