+++ backend/src/test/java/com/curevent/CureventApplicationTests.java
@@ -1,10 +1,10 @@
-package com.curevent.demo;
+package com.curevent;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class DemoApplicationTests {
+class CureventApplicationTests {
 
 	@Test
 	void contextLoads() {
